Revert "Event for pickup item"
This reverts commit 33b3ea164d8a5cc00e888b96da1f9c08c223439e.
